Rollback move to a new package
