+++ exampleApp/src/main/java/com/mercadolibre/android/ui/example/ui/widgets/dialog/DummyListDialog.java
@@ -7,7 +7,7 @@ import android.view.View;
 
 import com.mercadolibre.android.ui.example.R;
 import com.mercadolibre.android.ui.example.ui.SimpleAdapter;
-import com.mercadolibre.android.ui.widgets.dialogs.MeliDialog;
+import com.mercadolibre.android.ui.widgets.MeliDialog;
 
 public class DummyListDialog extends MeliDialog {
