+++ ui/src/main/java/com/mercadolibre/android/ui/widgets/MeliFullScreenDialog.java
@@ -1,4 +1,4 @@
-package com.mercadolibre.android.ui.widgets.dialogs;
+package com.mercadolibre.android.ui.widgets;
 
 import android.app.Dialog;
 import android.content.DialogInterface;
@@ -17,6 +17,7 @@ import android.view.ViewGroup;
 import android.widget.Button;
 
 import com.mercadolibre.android.ui.R;
+import com.mercadolibre.android.ui.widgets.MeliDialog;
 
 /**
  * Created by Mauro Rodriguez on 02/04/2019
